# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/accounts": {
      "post": {
        "operationId": "AccountsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountWithBalanceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "accounts"
        ]
      },
      "get": {
        "operationId": "AccountsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AccountWithBalanceDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "accounts"
        ]
      }
    },
    "/accounts/{id}": {
      "get": {
        "operationId": "AccountsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountWithBalanceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "accounts"
        ]
      },
      "patch": {
        "operationId": "AccountsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountWithBalanceDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "accounts"
        ]
      },
      "delete": {
        "operationId": "AccountsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The archived (soft-deleted) account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "accounts"
        ]
      }
    },
    "/categories": {
      "post": {
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "categories"
        ]
      },
      "get": {
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CategoryDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "categories"
        ]
      }
    },
...
    "/budgets": {
      "post": {
        "operationId": "BudgetsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBudgetDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetWithStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "budgets"
        ]
      },
      "get": {
        "operationId": "BudgetsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BudgetWithStatusDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "budgets"
        ]
      }
    },
    "/budgets/{id}": {
      "get": {
        "operationId": "BudgetsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetWithStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "budgets"
        ]
      },
      "patch": {
        "operationId": "BudgetsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBudgetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetWithStatusDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "budgets"
        ]
      },
      "delete": {
        "operationId": "BudgetsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The archived (soft-deleted) budget",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "budgets"
        ]
      }
    },
    "/analysis": {
      "get": {
        "operationId": "AnalysisController_getOverview",
        "parameters": [
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "default": "this_month",
              "type": "string",
              "enum": [
                "7d",
                "this_month",
                "last_month",
                "3m",
                "6m",
                "this_year",
                "custom"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "analysis"
        ]
      }
    },
    "/analysis/categories": {
      "get": {
        "operationId": "AnalysisController_getCategoryBreakdown",
        "parameters": [
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "default": "this_month",
              "type": "string",
              "enum": [
                "7d",
                "this_month",
                "last_month",
                "3m",
                "6m",
                "this_year",
                "custom"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryBreakdownDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "analysis"
        ]
      }
    },
    "/analysis/cash-flow": {
      "get": {
        "operationId": "AnalysisController_getCashFlow",
        "parameters": [
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "default": "this_month",
              "type": "string",
              "enum": [
                "7d",
                "this_month",
                "last_month",
                "3m",
                "6m",
                "this_year",
                "custom"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashFlowDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "analysis"
        ]
      }
    },
    "/analysis/accounts": {
      "get": {
        "operationId": "AnalysisController_getAccountBreakdown",
        "parameters": [
          {
            "name": "period",
            "required": false,
            "in": "query",
            "schema": {
              "default": "this_month",
              "type": "string",
              "enum": [
                "7d",
                "this_month",
                "last_month",
                "3m",
                "6m",
                "this_year",
                "custom"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z||([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountBreakdownDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "clerk": []
          }
        ],
        "tags": [
          "analysis"
        ]
      }
    }
  },
  "info": {
    "title": "Pocketly API",
    "description": "Core finance domain API for Pocketly: accounts, categories, transactions, budgets, and analysis.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "clerk": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Clerk session token"
      }
    },
    "schemas": {
...
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "bank",
              "cash",
              "savings",
              "upi",
              "credit_card",
              "wallet"
            ]
          },
          "icon": {
            "type": "string",
            "maxLength": 50
          },
          "initialBalance": {
            "default": 0,
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "currency": {
            "default": "INR",
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "AccountWithBalanceDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "userId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "name": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "type": {
            "type": "string",
            "enum": [
              "bank",
              "cash",
              "savings",
              "upi",
              "credit_card",
              "wallet"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "icon": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "initialBalance": {
            "type": "number",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "currency": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "ignored": {
            "type": "boolean",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "deletedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "createdAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "updatedAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "balance": {
            "type": "number",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        },
        "required": [
          "_id",
          "userId",
          "name",
          "type",
          "initialBalance",
          "currency",
          "ignored",
          "deletedAt",
          "createdAt",
          "updatedAt",
          "balance"
        ]
      },
      "UpdateAccountDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "bank",
              "cash",
              "savings",
              "upi",
              "credit_card",
              "wallet"
            ]
          },
          "icon": {
            "type": "string",
            "maxLength": 50
          },
          "initialBalance": {
            "default": 0,
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "currency": {
            "default": "INR",
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          },
          "ignored": {
            "type": "boolean"
          }
        }
      },
      "AccountDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "userId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "name": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "type": {
            "type": "string",
            "enum": [
              "bank",
              "cash",
              "savings",
              "upi",
              "credit_card",
              "wallet"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "icon": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "initialBalance": {
            "type": "number",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "currency": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "ignored": {
            "type": "boolean",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "deletedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "createdAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "updatedAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        },
        "required": [
          "_id",
          "userId",
          "name",
          "type",
          "initialBalance",
          "currency",
          "ignored",
          "deletedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ]
          },
          "icon": {
            "type": "string",
            "maxLength": 50
          },
          "color": {
            "type": "string",
            "maxLength": 20
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "CategoryDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "userId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "name": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "type": {
            "type": "string",
            "enum": [
              "income",
              "expense"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "icon": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "color": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "ignored": {
            "type": "boolean",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "deletedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "createdAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "updatedAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        },
        "required": [
          "_id",
          "userId",
          "name",
          "type",
          "ignored",
          "deletedAt",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateBudgetDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "amount": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "period": {
            "type": "string",
            "enum": [
              "weekly",
              "monthly",
              "yearly"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        },
        "required": [
          "categoryId",
          "amount",
          "period"
        ]
      },
      "BudgetWithStatusDto": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "userId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "categoryId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "amount": {
            "type": "number",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "period": {
            "type": "string",
            "enum": [
              "weekly",
              "monthly",
              "yearly"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "deletedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true,
            "x-nestjs_zod-empty-type": true
          },
          "syncVersion": {
            "type": "number",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "createdAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "updatedAt": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "spent": {
            "type": "number",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "remaining": {
            "type": "number",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "percentageUsed": {
            "type": "number",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "periodStart": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "periodEnd": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        },
        "required": [
          "_id",
          "userId",
          "categoryId",
          "amount",
          "period",
          "deletedAt",
          "syncVersion",
          "createdAt",
          "updatedAt",
          "spent",
          "remaining",
          "percentageUsed",
          "periodStart",
          "periodEnd"
        ]
      },
      "UpdateBudgetDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "string",
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "amount": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991,
            "x-nestjs_zod-uses-3-point-1-syntax": true
          },
          "period": {
            "type": "string",
            "enum": [
              "weekly",
              "monthly",
              "yearly"
            ],
            "x-nestjs_zod-uses-3-point-1-syntax": true
          }
        }
      },
...
    }
  }
}